added loginSubCompoenent to scanOrder to resolve history redirect complications
